fix(compiler-spec): extract shared semantic facts
