+++ packages/compiler/packages/compiler-spec/src/functionTypes.ts
@@ -1,3 +1,5 @@
+import type { LocalBinding, StackItem } from './semantic';
+
 export const SCALAR_TYPE_IDENTIFIERS = ['int', 'float', 'float64'] as const;
 
 export const POINTER_FUNCTION_TYPE_IDENTIFIERS = [
@@ -26,6 +28,112 @@ export type ScalarTypeIdentifier = (typeof SCALAR_TYPE_IDENTIFIERS)[number];
 export type PointerFunctionTypeIdentifier = (typeof POINTER_FUNCTION_TYPE_IDENTIFIERS)[number];
 export type FunctionTypeIdentifier = (typeof FUNCTION_TYPE_IDENTIFIERS)[number];
 export type FunctionValueType = FunctionTypeIdentifier;
+
+type PointerFunctionValueType = Extract<FunctionValueType, `${string}*`>;
+type PointerFunctionPointeeBaseType = NonNullable<LocalBinding['pointeeBaseType']>;
+
+const pointerFunctionValueTypes: ReadonlySet<string> = new Set(POINTER_FUNCTION_TYPE_IDENTIFIERS);
+const functionValueTypes: ReadonlySet<string> = new Set(FUNCTION_TYPE_IDENTIFIERS);
+
+export function isFunctionValueType(type: string): type is FunctionValueType {
+	return functionValueTypes.has(type);
+}
+
+export function isPointerFunctionValueType(type: FunctionValueType): type is PointerFunctionValueType {
+	return pointerFunctionValueTypes.has(type);
+}
+
+function getPointerParts(type: PointerFunctionValueType) {
+	const pointerDepth = type.endsWith('**') ? 2 : 1;
+	const baseType = type.replace(/\*+$/, '') as PointerFunctionPointeeBaseType;
+	return { baseType, pointerDepth };
+}
+
+export function functionValueTypeToLocalBinding(type: FunctionValueType, index: number): LocalBinding {
+	if (type === 'int') {
+		return { isInteger: true, index };
+	}
+
+	if (type === 'float') {
+		return { isInteger: false, index };
+	}
+
+	if (type === 'float64') {
+		return { isInteger: false, isFloat64: true, index };
+	}
+
+	if (!isPointerFunctionValueType(type)) {
+		return { isInteger: false, index };
+	}
+
+	const { baseType, pointerDepth } = getPointerParts(type);
+	return {
+		isInteger: true,
+		pointeeBaseType: baseType,
+		pointerDepth,
+		index,
+	};
+}
+
+function localBindingToStackItem(binding: LocalBinding): StackItem {
+	if (binding.pointeeBaseType) {
+		return {
+			kind: 'address',
+			valueType: 'int',
+			address: {
+				memoryIndex: binding.pointeeMemoryIndex ?? 0,
+				...(binding.pointeeMemoryRegionName ? { memoryRegionName: binding.pointeeMemoryRegionName } : {}),
+			},
+			pointsTo: {
+				baseType: binding.pointeeBaseType,
+				memoryIndex: binding.pointeeMemoryIndex ?? 0,
+				...(binding.pointeeMemoryRegionName ? { memoryRegionName: binding.pointeeMemoryRegionName } : {}),
+				pointerDepth: binding.pointerDepth,
+			},
+		};
+	}
+
+	return {
+		kind: 'value',
+		valueType: binding.isInteger ? 'int' : binding.isFloat64 ? 'float64' : 'float',
+	};
+}
+
+export function functionValueTypeToStackItem(type: FunctionValueType): StackItem {
+	return localBindingToStackItem(functionValueTypeToLocalBinding(type, 0));
+}
+
+export function stackItemMatchesFunctionValueType(stackItem: StackItem, type: FunctionValueType): boolean {
+	if (type === 'int') {
+		return stackItem.valueType === 'int';
+	}
+
+	if (type === 'float') {
+		return stackItem.valueType === 'float';
+	}
+
+	if (type === 'float64') {
+		return stackItem.valueType === 'float64';
+	}
+
+	if (stackItem.valueType !== 'int') {
+		return false;
+	}
+
+	if (stackItem.kind !== 'address' || !stackItem.pointsTo) {
+		return true;
+	}
+
+	const expected = functionValueTypeToStackItem(type);
+	if (expected.kind !== 'address' || !expected.pointsTo) {
+		return true;
+	}
+
+	return (
+		stackItem.pointsTo.baseType === expected.pointsTo.baseType &&
+		stackItem.pointsTo.pointerDepth === expected.pointsTo.pointerDepth
+	);
+}
 
 /** Default WebAssembly import module namespace for 8f4e host-provided imports. */
 export const DEFAULT_HOST_IMPORT_MODULE_NAME = 'host';
