+++ packages/compiler/packages/compiler-spec/src/index.ts
@@ -12,7 +12,10 @@ export * from './instructionSpecTypes';
 export * from './instructionSpecUtils';
 export * from './instructions';
 export * from './memory';
+export * from './memoryData';
+export * from './memoryRegions';
 export * from './options';
+export * from './paramShape';
 export * from './runtime';
 export * from './semantic';
 export * from './testModule';
