+++ packages/compiler/packages/compiler-spec/src/memory.ts
@@ -53,6 +53,26 @@ export const memoryDeclarationInstructions = [
 	...arrayMemoryDeclarationInstructions,
 ] as readonly MemoryDeclarationInstruction[];
 
+const scalarMemoryDeclarationInstructionSet: ReadonlySet<string> = new Set(scalarMemoryDeclarationInstructions);
+const arrayMemoryDeclarationInstructionSet: ReadonlySet<string> = new Set(arrayMemoryDeclarationInstructions);
+const memoryDeclarationInstructionSet: ReadonlySet<string> = new Set(memoryDeclarationInstructions);
+
+export function isScalarMemoryDeclarationInstructionName(
+	instruction: string
+): instruction is ScalarMemoryDeclarationInstruction {
+	return scalarMemoryDeclarationInstructionSet.has(instruction);
+}
+
+export function isArrayMemoryDeclarationInstructionName(
+	instruction: string
+): instruction is ArrayDeclarationInstruction {
+	return arrayMemoryDeclarationInstructionSet.has(instruction);
+}
+
+export function isMemoryDeclarationInstructionName(instruction: string): instruction is MemoryDeclarationInstruction {
+	return memoryDeclarationInstructionSet.has(instruction);
+}
+
 export type BaseMemoryType = 'int' | 'int8' | 'int16' | 'float' | 'float64';
 type ReservedUnsignedBaseMemoryType = 'int8u' | 'int16u';
 export type PointeeBaseType = BaseMemoryType | ReservedUnsignedBaseMemoryType;
