+++ packages/compiler/packages/memory-reference-inliner/src/addressValues.ts
@@ -1,7 +1,6 @@
 import type { Const, PlannedMemoryDeclaration } from '@8f4e/compiler-spec';
-import { GLOBAL_ALIGNMENT_BOUNDARY } from '@8f4e/compiler-spec';
+import { GLOBAL_ALIGNMENT_BOUNDARY, getMemoryRegionFields } from '@8f4e/compiler-spec';
 import { getEndByteAddress } from './layoutAddresses';
-import { getMemoryRegionFields } from './memoryRegions';
 
 type AddressableMemoryDeclaration = Pick<
 	PlannedMemoryDeclaration,
