+++ packages/compiler/packages/stack-analyzer/src/instructionAnalyzers/call.ts
@@ -7,8 +7,7 @@ import type {
 	Stack,
 	StackItem,
 } from '@8f4e/compiler-spec';
-import { createFunctionId, ErrorCode, getError } from '@8f4e/compiler-spec';
-import { functionValueTypeToStackItem } from '../utils/functionValueType';
+import { createFunctionId, ErrorCode, functionValueTypeToStackItem, getError } from '@8f4e/compiler-spec';
 import { analyzePush } from './push';
 import { consume, produce } from './stack';
