--- packages/compiler/packages/stack-analyzer/src/utils/memoryRegionFields.ts
@@ -1,8 +1,0 @@
-import type { MemoryRegionIdentity } from '@8f4e/compiler-spec';
-
-export function getMemoryRegionFields(memoryIndex: number, memoryRegionName?: string): MemoryRegionIdentity {
-	return {
-		memoryIndex,
-		...(memoryRegionName ? { memoryRegionName } : {}),
-	};
-}
