+++ packages/compiler/packages/stack-analyzer/src/utils/stackAddressMetadata.ts
@@ -1,5 +1,5 @@
 import type { MemoryAddressRange, StackAddress, StackItem } from '@8f4e/compiler-spec';
-import { getMemoryRegionFields } from './memoryRegionFields';
+import { getMemoryRegionFields } from '@8f4e/compiler-spec';
 
 /** Shifts a proven safe range forward by a known byte offset, dropping it when unsafe. */
 function shiftSafeRange(safeRange: MemoryAddressRange, byteOffset: number): MemoryAddressRange | undefined {
