+++ packages/compiler/src/instructionCompilers/param.ts
@@ -1,5 +1,5 @@
 import type { FunctionCodegenContext, FunctionValueType, InstructionCompiler, ParamLine } from '@8f4e/compiler-spec';
-import { functionValueTypeToLocalBinding } from '../utils/functionValueType';
+import { functionValueTypeToLocalBinding } from '@8f4e/compiler-spec';
 
 /**
  * Instruction compiler for `param`.
