+++ packages/compiler/src/instructionCompilers/push/shared.ts
@@ -1,5 +1,10 @@
-import type { CodegenContext } from '@8f4e/compiler-spec';
-import { WORD_MEMORY_ACCESS_WIDTH } from '@8f4e/compiler-spec';
+import type { CodegenContext, PointerMetadata } from '@8f4e/compiler-spec';
+import {
+	getDereferencedValueKindFromMetadata,
+	getDereferencedValueWordSizeFromMetadata,
+	getPointerDepthFromMetadata,
+	WORD_MEMORY_ACCESS_WIDTH,
+} from '@8f4e/compiler-spec';
 import {
 	f32const,
 	f32load,
@@ -13,12 +18,6 @@ import {
 	i32load16u,
 	localGet,
 } from '@8f4e/compiler-wasm-utils';
-import type { PointerMetadata } from '../../utils/memoryData';
-import {
-	getDereferencedValueKindFromMetadata,
-	getDereferencedValueWordSizeFromMetadata,
-	getPointerDepthFromMetadata,
-} from '../../utils/memoryData';
 import { type PushValueKind, valueKindToWasmType } from '../../utils/pushValueKind';
 import { guardedAddressOperation } from '../utils/memoryAccessGuard';
